+++ @xen-orchestra/rest-api/src/vms/vm.controller.mts
@@ -8,7 +8,7 @@ import type { XapiStatsGranularity, XapiVmStats, XoVm } from '@vates/types'
 import { partialVms, vm, vmIds } from '../open-api/oa-examples/vm.oa-example.mjs'
 import { RestApi } from '../rest-api/rest-api.mjs'
 import type { Unbrand, WithHref } from '../helpers/helper.type.mjs'
-import { vmStatsExample } from '../open-api/examples/vm.examples.mjs'
+import { vmStatsExample } from '../open-api/oa-examples/vm.oa-example.mjs'
 import { XapiXoController } from '../abstract-classes/xapi-xo-controller.mjs'
 
 @Route('vms')
